requirements.txt
altgraph==0.16.1
Django==2.1.2
djangorestframework==3.8.2
future==0.17.1
macholib==1.11
pefile==2018.8.8
psycopg2-binary==2.7.6.1
PyInstaller==3.4
pytz==2018.5
pywin32-ctypes==0.2.0
